refactor(types): move the modify contract into microsandbox-types
Relocate the serializable modify request/response types — the patch,
plan, planned changes, dispositions, policies, secret patch types,
and resize status — from the SDK into a new modify module in the
shared microsandbox-types crate, with ts derives matching the domain
types. They are already the wire contract every language binding
speaks as JSON; owning them in the shared-contract crate means a
future cloud backend can reuse the same shapes for its client/server
protocol instead of promoting them later under stability pressure.

Behavior stays where it was: the SDK keeps the modification builder,
classification, and apply logic, and re-exports the moved types on
their existing public paths, so nothing changes for SDK consumers or
the bindings.
